+++ src/main/java/de/samply/converter/Converter.java
@@ -5,4 +5,6 @@ import reactor.core.publisher.Flux;
 
 public interface Converter<I,O> {
   Flux<O> convert (Flux<I> input, ConversionTemplate template);
+  Format getInputFormat();
+  Format getOutputFormat();
 }