+++ src/main/java/de/samply/converter/ConverterGroup.java
@@ -5,7 +5,8 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 import reactor.core.publisher.Flux;
 
-public class ConverterGroup<I,O> implements Converter<I,O>{
+public class ConverterGroup<I, O> implements Converter<I, O> {
+
   private List<Converter> converters;
 
   public ConverterGroup(List<Converter> converters) {
@@ -17,6 +18,18 @@ public class ConverterGroup<I,O> implements Converter<I,O>{
     AtomicReference<Flux> tempFlux = new AtomicReference<>(input);
     converters.forEach(converter -> tempFlux.set(converter.convert(tempFlux.get(), template)));
     return tempFlux.get();
+  }
+
+  @Override
+  public Format getInputFormat() {
+    return (converters != null && converters.size() > 0) ? converters.get(0).getInputFormat()
+        : null;
+  }
+
+  @Override
+  public Format getOutputFormat() {
+    return (converters != null && converters.size() > 0) ? converters.get(converters.size() - 1)
+        .getOutputFormat() : null;
   }
 
 }