+++ CHANGELOG.md
@@ -16,3 +16,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Converter Pattern
 - Converter Manager
 - Converter Template and Conversion Template
+- Input and Output Format in Converter